Made desktop-specific states, transitions and signals (like TERMINATE) condition compiled.
