+++ src_c/extern_sm/bsp_desktop.h
@@ -13,4 +13,13 @@ Module to replace for various platforms.
 #include "vibro_desktop.h"
 #include "RGB_desktop.h"
 
+#define DESKTOP
+
+#ifdef DESKTOP
+    #include "generation_light.h"
+    enum DesktopSignal {
+        TERMINATE_SIG = LAST_USER_SIG + 1 /* terminate the application */
+    };
+#endif
+
 #endif
