+++ src_c/extern_sm/service.c
@@ -1,0 +1,32 @@
+#include "service.h"
+
+const KeyStroke KeyStrokes[]= {
+{    CHARGE_SIG,         "CHARGE",           '`'    },
+{    THROW_SIG,          "THROW",            '1'    },
+{    PUNCH_SIG,          "PUNCH",            '2'    },
+{    LIFT_SIG,           "LIFT",             '3'    },
+{    WARP_SIG,           "WARP",             '4'    },
+{    BARRIER_SIG,        "BARRIER",          '5'    },
+{    CLEANSE_SIG,        "CLEANSE",          '6'    },
+{    SINGULAR_SIG,       "SINGULAR",         '7'    },
+{    SONG_SIG,           "SONG",             '8'    },
+{    RELEASE_SIG,        "RELEASE",          '9'    },
+{    PWR_RELEASE_SIG,    "PWR_RELEASE",      '0'    },
+
+{    PUNCH_PILL_SIG,             "PUNCH_PILL",               'p'    },
+{    PUNCH_PWR_PILL_SIG,         "PUNCH_PWR_PILL",           'P'    },
+{    LIFT_PILL_SIG,              "LIFT_PILL",                'l'    },
+{    LIFT_PWR_PILL_SIG,          "LIFT_PWR_PILL",            'L'    },
+{    WARP_PILL_SIG,              "WARP_PILL",                'w'    },
+{    WARP_PWR_PILL_SIG,          "WARP_PWR_PILL",            'W'    },
+{    CLEANSE_PILL_SIG,           "CLEANSE_PILL",             'c'    },
+{    BARRIER_PILL_SIG,           "BARRIER_PILL",             'b'    },
+{    SINGULARITY_PILL_SIG,       "SINGULARITY_PILL",         'S'    },
+{    SONG_PILL_SIG,              "SONG_PILL",                's'    },
+{    DEFAULT_PILL_SIG,           "DEFAULT_PILL",             'd'    },
+{    MAX_PILL_SIG,               "MAX_PILL",                 'M'    },
+
+                
+{    TERMINATE_SIG,          "TERMINATE",            0x1B   }
+
+};
