+++ src_c/extern_sm/vibro_desktop.c
@@ -1,0 +1,4 @@
+#include "vibro_desktop.h"
+void vibro(uint8_t Power) {
+    printf("vibro(%u);", Power);
+}
