+++ src_c/extern_sm/vibro_desktop.h
@@ -4,10 +4,9 @@ Module to emulate vibro functions.
 #ifndef vibro_desktop_h
 #define vibro_desktop_h
 #include <stdint.h>
+#include <stdio.h>
 
 #define OFF 0
-void vibro(uint8_t Power) {
-    printf("vibro(%u);", Power);
-}
+void vibro(uint8_t Power);
 
 #endif
